Decouple comments badge from view (#200974)
* Decouple comments badge from view
Fixes #200972

* Dispose disposable
