+++ src/vs/workbench/contrib/comments/browser/commentsView.ts
@@ -10,7 +10,7 @@ import { basename } from 'vs/base/common/resources';
 import { isCodeEditor, isDiffEditor } from 'vs/editor/browser/editorBrowser';
 import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
-import { CommentNode, CommentsModel, ResourceWithCommentThreads, ICommentThreadChangedEvent } from 'vs/workbench/contrib/comments/common/commentModel';
+import { CommentNode, ResourceWithCommentThreads, ICommentThreadChangedEvent } from 'vs/workbench/contrib/comments/common/commentModel';
 import { IWorkspaceCommentThreadsEvent, ICommentService } from 'vs/workbench/contrib/comments/browser/commentService';
 import { IEditorService, ACTIVE_GROUP, SIDE_GROUP } from 'vs/workbench/services/editor/common/editorService';
 import { CommandsRegistry } from 'vs/platform/commands/common/commands';
@@ -35,20 +35,19 @@ import { CommentsFilters, CommentsFiltersChangeEvent } from 'vs/workbench/contri
 import { Memento, MementoObject } from 'vs/workbench/common/memento';
 import { IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
 import { FilterOptions } from 'vs/workbench/contrib/comments/browser/commentsFilterOptions';
-import { IActivityService, NumberBadge } from 'vs/workbench/services/activity/common/activity';
 import { CommentThreadState } from 'vs/editor/common/languages';
-import { IDisposable, MutableDisposable } from 'vs/base/common/lifecycle';
 import { ITreeElement } from 'vs/base/browser/ui/tree/tree';
 import { Iterable } from 'vs/base/common/iterator';
 import { CommentController } from 'vs/workbench/contrib/comments/browser/commentsController';
 import { Range } from 'vs/editor/common/core/range';
 import { registerNavigableContainer } from 'vs/workbench/browser/actions/widgetNavigationCommands';
+import { CommentsModel, ICommentsModel } from 'vs/workbench/contrib/comments/browser/commentsModel';
 
 const CONTEXT_KEY_HAS_COMMENTS = new RawContextKey<boolean>('commentsView.hasComments', false);
 const CONTEXT_KEY_SOME_COMMENTS_EXPANDED = new RawContextKey<boolean>('commentsView.someCommentsExpanded', false);
 const VIEW_STORAGE_ID = 'commentsViewState';
 
-function createResourceCommentsIterator(model: CommentsModel): Iterable<ITreeElement<ResourceWithCommentThreads | CommentNode>> {
+function createResourceCommentsIterator(model: ICommentsModel): Iterable<ITreeElement<ResourceWithCommentThreads | CommentNode>> {
 	return Iterable.map(model.resourceCommentThreads, m => {
 		const CommentNodeIt = Iterable.from(m.commentThreads);
 		const children = Iterable.map(CommentNodeIt, r => ({ element: r }));
@@ -62,14 +61,11 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 	private tree: CommentsList | undefined;
 	private treeContainer!: HTMLElement;
 	private messageBoxContainer!: HTMLElement;
-	private commentsModel!: CommentsModel;
 	private totalComments: number = 0;
-	private totalUnresolved = 0;
 	private readonly hasCommentsContextKey: IContextKey<boolean>;
 	private readonly someCommentsExpandedContextKey: IContextKey<boolean>;
 	private readonly filter: Filter;
 	readonly filters: CommentsFilters;
-	private readonly activity = this._register(new MutableDisposable<IDisposable>());
 
 	private currentHeight = 0;
 	private currentWidth = 0;
@@ -93,7 +89,6 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 		@ICommentService private readonly commentService: ICommentService,
 		@ITelemetryService telemetryService: ITelemetryService,
 		@IUriIdentityService private readonly uriIdentityService: IUriIdentityService,
-		@IActivityService private readonly activityService: IActivityService,
 		@IStorageService storageService: IStorageService
 	) {
 		const stateMemento = new Memento(VIEW_STORAGE_ID, storageService);
@@ -125,16 +120,6 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 			}
 		}));
 		this._register(this.filterWidget.onDidChangeFilterText(() => this.updateFilter()));
-	}
-
-	private updateBadge(unresolved: number) {
-		if (unresolved === this.totalUnresolved) {
-			return;
-		}
-
-		this.totalUnresolved = unresolved;
-		const message = nls.localize('totalUnresolvedComments', '{0} Unresolved Comments', this.totalUnresolved);
-		this.activity.value = this.activityService.showViewActivity(this.id, { badge: new NumberBadge(this.totalUnresolved, () => message) });
 	}
 
 	override saveState(): void {
@@ -201,7 +186,6 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 
 		this.treeContainer = dom.append(domContainer, dom.$('.tree-container'));
 		this.treeContainer.classList.add('file-icon-themable-tree', 'show-file-icons');
-		this.commentsModel = new CommentsModel();
 
 		this.cachedFilterStats = undefined;
 		this.createTree();
@@ -232,7 +216,7 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 			return;
 		}
 
-		if (!this.commentsModel.hasCommentThreads() && this.messageBoxContainer) {
+		if (!this.commentService.commentsModel.hasCommentThreads() && this.messageBoxContainer) {
 			this.messageBoxContainer.focus();
 		} else if (this.tree) {
 			this.tree.domFocus();
@@ -267,9 +251,9 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 	}
 
 	private async renderComments(): Promise<void> {
-		this.treeContainer.classList.toggle('hidden', !this.commentsModel.hasCommentThreads());
+		this.treeContainer.classList.toggle('hidden', !this.commentService.commentsModel.hasCommentThreads());
 		this.renderMessage();
-		await this.tree?.setChildren(null, createResourceCommentsIterator(this.commentsModel));
+		await this.tree?.setChildren(null, createResourceCommentsIterator(this.commentService.commentsModel));
 	}
 
 	public collapseAll() {
@@ -311,8 +295,8 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 	}
 
 	private renderMessage(): void {
-		this.messageBoxContainer.textContent = this.commentsModel.getMessage();
-		this.messageBoxContainer.classList.toggle('hidden', this.commentsModel.hasCommentThreads());
+		this.messageBoxContainer.textContent = this.commentService.commentsModel.getMessage();
+		this.messageBoxContainer.classList.toggle('hidden', this.commentService.commentsModel.hasCommentThreads());
 	}
 
 	private createTree(): void {
@@ -437,15 +421,15 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 			return;
 		}
 		if (this.isVisible()) {
-			this.hasCommentsContextKey.set(this.commentsModel.hasCommentThreads());
+			this.hasCommentsContextKey.set(this.commentService.commentsModel.hasCommentThreads());
 
-			this.treeContainer.classList.toggle('hidden', !this.commentsModel.hasCommentThreads());
+			this.treeContainer.classList.toggle('hidden', !this.commentService.commentsModel.hasCommentThreads());
 			this.cachedFilterStats = undefined;
 			this.renderMessage();
-			this.tree?.setChildren(null, createResourceCommentsIterator(this.commentsModel));
+			this.tree?.setChildren(null, createResourceCommentsIterator(this.commentService.commentsModel));
 
-			if (this.tree.getSelection().length === 0 && this.commentsModel.hasCommentThreads()) {
-				const firstComment = this.commentsModel.resourceCommentThreads[0].commentThreads[0];
+			if (this.tree.getSelection().length === 0 && this.commentService.commentsModel.hasCommentThreads()) {
+				const firstComment = this.commentService.commentsModel.resourceCommentThreads[0].commentThreads[0];
 				if (firstComment) {
 					this.tree.setFocus([firstComment]);
 					this.tree.setSelection([firstComment]);
@@ -456,7 +440,6 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 
 	private onAllCommentsChanged(e: IWorkspaceCommentThreadsEvent): void {
 		this.cachedFilterStats = undefined;
-		this.commentsModel.setCommentThreads(e.ownerId, e.ownerLabel, e.commentThreads);
 		this.totalComments += e.commentThreads.length;
 
 		let unresolved = 0;
@@ -465,37 +448,28 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 				unresolved++;
 			}
 		}
-		this.updateBadge(unresolved);
-
 		this.refresh();
 	}
 
 	private onCommentsUpdated(e: ICommentThreadChangedEvent): void {
 		this.cachedFilterStats = undefined;
-		const didUpdate = this.commentsModel.updateCommentThreads(e);
 
 		this.totalComments += e.added.length;
 		this.totalComments -= e.removed.length;
 
 		let unresolved = 0;
-		for (const resource of this.commentsModel.resourceCommentThreads) {
+		for (const resource of this.commentService.commentsModel.resourceCommentThreads) {
 			for (const thread of resource.commentThreads) {
 				if (thread.threadState === CommentThreadState.Unresolved) {
 					unresolved++;
 				}
 			}
 		}
-		this.updateBadge(unresolved);
-
-		if (didUpdate) {
-			this.refresh();
-		}
+		this.refresh();
 	}
 
 	private onDataProviderDeleted(owner: string | undefined): void {
 		this.cachedFilterStats = undefined;
-		this.commentsModel.deleteCommentsByOwner(owner);
-
 		this.totalComments = 0;
 		this.refresh();
 	}
